back/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-login==0.6.2
sqlalchemy==2.0.21
python-dotenv==1.0.0
requests==2.31.0
playwright==1.39.0
langchain==0.0.350
langchain-openai==0.0.5
openai>=1.10.0,<2.0.0

PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.3
numpy==1.25.2
celery==5.3.4
redis==5.0.1

